+++ tests/index.rs
@@ -168,9 +168,8 @@ async fn test_files() {
     echo_server.stop(true).await
 }
 
-#[actix_web::test]
-async fn test_file_upload() -> actix_web::Result<()> {
-    let req = get_request_to("/tests/upload_file_test.sql")
+async fn test_file_upload(target: &str) -> actix_web::Result<()> {
+    let req = get_request_to(target)
         .await?
         .insert_header(("content-type", "multipart/form-data; boundary=1234567890"))
         .set_payload(
@@ -192,6 +191,16 @@ async fn test_file_upload() -> actix_web::Result<()> {
         "{body_str}\nexpected to contain: Hello, world!"
     );
     Ok(())
+}
+
+#[actix_web::test]
+async fn test_file_upload_direct() -> actix_web::Result<()> {
+    test_file_upload("/tests/upload_file_test.sql").await
+}
+
+#[actix_web::test]
+async fn test_file_upload_through_runsql() -> actix_web::Result<()> {
+    test_file_upload("/tests/upload_file_runsql_test.sql").await
 }
 
 #[actix_web::test]